requirements.txt
flask==2.0.1
werkzeug==2.0.3
requests==2.31.0
icalendar==5.0.7
python-dateutil==2.8.2
pytz==2023.3
firebase-admin==6.2.0
google-cloud-firestore==2.11.1
python-dotenv==1.0.0
gunicorn==21.2.0
